refactor: Project Filter Media Query 로직을 page에서 별도 컴포넌트로 이동
