+++ src/views/ProjectPage/components/index.ts
@@ -1,6 +1,5 @@
 export * from './common/EmptyContent';
-export * from './filter/DesktopFilter';
-export * from './filter/MobileFilter';
+export * from './filter';
 export * from './project/ProjectCard';
 export * from './project/ProjectEnrollSection';
 export * from './project/ProjectList';
